+++ CHANGELOG.md
@@ -33,6 +33,7 @@
   need to pay back an existing output with a slight modification.
 - A new tweak `modifySpendRedeemersOfTypeTweak` to apply an optional
   modification of all redeemers of a certain type within the skeleton inputs.
+- New setters for the various outputs fields.
 
 ### Removed
 
@@ -64,7 +65,8 @@
 - Dependency to cardano-api bumped to 8.46.
 - The whole testing API has been revamped
 - File `AddInputsAndOutputs.hs` has been split into `Inputs.hs`, `Outputs.hs`
-  and `Mint.hs`.
+  and `Mint.hs`. File `TamperDatum.hs` has been updated and integrated into
+  `Output.hs`.
 
 ### Fixed